requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.4
Django==4.2.23
django-bootstrap-v5==1.0.11
django-ckeditor==6.7.3
django-js-asset==3.1.2
pillow==11.3.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
